backend/requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
confusable-homoglyphs==3.2.0
Django==4.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
